// backup/features/stores/types/types.ts
export type ProcessedShippingMethod = {
    provider: string
    min_order_amount: number | null
    free_shipping_min: number | null
    eta_minutes: number | null
    delivery_price: number | null
    active: boolean
}